Move mapToBLTypes method to JSONUtil for reuse
